refactor(manager): 변수명 변경 및 디버깅코드 제거
