+++ apps/manager/src/api/mutations/useDeleteTimeline.ts
@@ -6,18 +6,18 @@ type Request = {
   timelineId: number;
 };
 
-export const deleteTimelines = ({ gameId, timelineId }: Request) => {
+export const deleteTimeline = ({ gameId, timelineId }: Request) => {
   return fetcher.delete<void>(`games/${gameId}/timelines/${timelineId}`, {
     json: null,
   });
 };
 
-export const useDeleteTimelines = ({ gameId }: { gameId: number }) => {
+export const useDeleteTimeline = ({ gameId }: { gameId: number }) => {
   const qc = useQueryClient();
 
   return useMutation({
     mutationKey: queryKeys.games.timeline({ gameId }).queryKey,
-    mutationFn: deleteTimelines,
+    mutationFn: deleteTimeline,
     onSuccess: async () => {
       await qc.invalidateQueries({
         queryKey: queryKeys.games.timeline({ gameId }).queryKey,
